backend/requirements.txt
accelerate==1.0.1
aniso8601==9.0.1
asttokens==2.4.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
diffusers==0.30.3
distlib==0.3.9
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
fsspec==2024.9.0
huggingface-hub==0.25.2
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.1
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.1
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
packaging==24.1
parso==0.8.4
pbr==6.1.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
setuptools==75.2.0
six==1.16.0
stack-data==0.6.3
stevedore==5.3.0
sympy==1.13.3
tokenizers==0.20.1
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.2
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.26.6
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.1
wcwidth==0.2.13
Werkzeug==3.0.4
zipp==3.20.2
